feat+refactor+fix: the grand path rasterization refactor.
this pretty much implements how path rasterization is supposed to
work for shapes. most of the code has been inspired by as3swf, the
corresponding blog post and the blog post by m. swanson on msdn.

this also partially fixes how shapes are supposed to be laid out.

i think the current way we do things is a bit off and we should be
rasterizing the paths right as we draw them in the future, but meh,
this is a toy project after all.