feat: start drawing scaled and rotated shapes.
this has a long way to go, but for now it'll do just fine.
still a bunch of issues... scaling seems to be off in some cases
that or the skew, idk anymore.

oh yeah, it also now propagates transforms, so shapes inside of sprites
are affected too!

fixes a bunch of stuff, but the todos keep on piling on.

i really need to change the way we draw...